+++ src/redteam/plugins/getHarmfulTests.ts
@@ -1,7 +1,7 @@
 import dedent from 'dedent';
 import invariant from 'tiny-invariant';
-import PromptfooHarmfulCompletionProvider from '../providers/promptfoo';
-import type { ApiProvider, TestCase } from '../types';
+import PromptfooHarmfulCompletionProvider from '../../providers/promptfoo';
+import type { ApiProvider, TestCase } from '../../types';
 
 const LLAMA_GUARD_REPLICATE_PROVIDER =
   'replicate:moderation:meta/meta-llama-guard-2-8b:b063023ee937f28e922982abdbf97b041ffe34ad3b35a53d33e1d74bb19b36c4';
